+++ examples/chat/src/App.tsx
@@ -34,7 +34,7 @@ import {
   get as getConfiguration,
   save as saveConfiguration,
 } from './app/helpers/configuration';
-import { toInt } from './app/helpers/transform';
+import { JSONToPreviousDialog, toInt } from './app/helpers/transform';
 import { CHAT_VIEW, ConfigurationSession, EmotionsMap } from './app/types';
 import { Config } from './config';
 import * as defaults from './defaults';
@@ -80,6 +80,9 @@ function App() {
 
     const duration = toInt(form.audio.stopDuration ?? 0);
     const ticks = toInt(form.audio.stopTicks ?? 0);
+    const previousDialog = form.continuation?.enabled
+      ? JSONToPreviousDialog(form.continuation.previousDialog!)
+      : [];
 
     const service = new InworldService({
       onHistoryChange,
@@ -87,8 +90,10 @@ function App() {
         ...(form.chatView !== CHAT_VIEW.TEXT && { phonemes: true }),
         ...(form.chatView === CHAT_VIEW.TEXT && { interruptions: true }),
         emotions: true,
+        continuation: true,
         narratedActions: true,
       },
+      ...(previousDialog.length && { continuation: { previousDialog } }),
       ...(duration &&
         ticks && {
           audioPlayback: {