+++ src/connection/web-socket.connection.ts
@@ -3,7 +3,6 @@ import {
   Awaitable,
   InternalClientConfiguration,
   SessionToken,
-  VoidFn,
 } from '../common/data_structures';
 import { InworldPacket } from '../entities/inworld_packet.entity';
 
@@ -12,16 +11,16 @@ const SESSION_PATH = '/v1/session/default';
 interface SessionProps {
   config: InternalClientConfiguration;
   session: SessionToken;
-  onDisconnect?: VoidFn;
-  onError?: (err: Event | Error) => void;
+  onDisconnect?: () => Awaitable<void>;
+  onError?: (err: Event | Error) => Awaitable<void>;
   onMessage?: (packet: ProtoPacket) => Awaitable<void>;
-  onReady?: VoidFn;
+  onReady?: () => Awaitable<void>;
 }
 interface ConnectionProps {
   config?: InternalClientConfiguration;
-  onDisconnect?: VoidFn;
+  onDisconnect?: () => Awaitable<void>;
   onReady?: () => Awaitable<void>;
-  onError?: (err: Event | Error) => void;
+  onError?: (err: Event | Error) => Awaitable<void>;
   onMessage?: (packet: ProtoPacket) => Awaitable<void>;
 }
 
@@ -104,7 +103,7 @@ export class WebSocketConnection<
     // So put packets to queue and send them `onReady` event.
     if (this.isActive()) {
       const packet = item.getPacket();
-      const inworldPacket = this.convertPacketFromProto(item.getPacket());
+      const inworldPacket = this.convertPacketFromProto(packet);
       await item.beforeWriting?.(inworldPacket);
       this.ws.send(JSON.stringify(packet));
       item.afterWriting?.(inworldPacket);